requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
plotly
google-generativeai
python-dotenv
langchain
python-decouple
